refactor: base DatesModelMixin created for dates
