+++ .github/workflows/action.yaml
@@ -2,8 +2,10 @@ name: Build and deploy workflow
 on:
   push:
     branches:
-      - main
-      - goals_app
+#      - 'goals_app'
+  pull_request:
+    branches:
+      - 'main'
 jobs:
   build_and_push:
     runs-on: ubuntu-latest
