+++ todolist/goals/models/category.py
@@ -1,10 +1,10 @@
 from django.db import models
-from django.utils import timezone
 
 from core.models import User
+from .dates_mixin import DatesModelMixin
 
 
-class Category(models.Model):
+class Category(DatesModelMixin):
     class Meta:
         verbose_name = "Категория"
         verbose_name_plural = "Категории"
@@ -12,11 +12,3 @@ class Category(models.Model):
     title = models.CharField(verbose_name="Название", max_length=255)
     user = models.ForeignKey(User, verbose_name="Автор", on_delete=models.PROTECT)
     is_deleted = models.BooleanField(verbose_name="Удалена", default=False)
-    created = models.DateTimeField(verbose_name="Дата создания")
-    updated = models.DateTimeField(verbose_name="Дата последнего обновления")
-
-    def save(self, *args, **kwargs):
-        if not self.id:
-            self.created = timezone.now()
-        self.updated = timezone.now()
-        return super().save(*args, **kwargs)
