+++ todolist/goals/models/comment.py
@@ -1,11 +1,11 @@
 from django.db import models
-from django.utils import timezone
 
 from core.models import User
 from .goal import Goal
+from .dates_mixin import DatesModelMixin
 
 
-class Comment(models.Model):
+class Comment(DatesModelMixin):
     class Meta:
         verbose_name = "Комментарий"
         verbose_name_plural = "Комментарии"
@@ -13,11 +13,3 @@ class Comment(models.Model):
     user = models.ForeignKey(User, verbose_name="Автор", on_delete=models.PROTECT)
     goal = models.ForeignKey(Goal, verbose_name="Цель", on_delete=models.CASCADE)
     text = models.CharField(verbose_name="Текст", max_length=1200)
-    created = models.DateTimeField(verbose_name="Дата создания")
-    updated = models.DateTimeField(verbose_name="Дата последнего обновления")
-
-    def save(self, *args, **kwargs):
-        if not self.id:
-            self.created = timezone.now()
-        self.updated = timezone.now()
-        return super().save(*args, **kwargs)
