+++ todolist/goals/models/goal.py
@@ -1,8 +1,8 @@
 from django.db import models
-from django.utils import timezone
 
 from core.models import User
 from goals.models import Category
+from .dates_mixin import DatesModelMixin
 
 
 class Status(models.IntegerChoices):
@@ -19,15 +19,13 @@ class Priority(models.IntegerChoices):
     critical = 4, "Критический"
 
 
-class Goal(models.Model):
+class Goal(DatesModelMixin):
     class Meta:
         verbose_name = "Цель"
         verbose_name_plural = "Цели"
 
     title = models.CharField(verbose_name="Заголовок", max_length=255)
     description = models.CharField(verbose_name="Описание", max_length=500)
-    created = models.DateTimeField(verbose_name="Дата создания")
-    updated = models.DateTimeField(verbose_name="Дата последнего обновления")
     due_date = models.DateTimeField(verbose_name="Дата выполнения")
     status = models.PositiveSmallIntegerField(
         verbose_name="Статус", choices=Status.choices, default=Status.to_do
@@ -38,9 +36,3 @@ class Goal(models.Model):
     user = models.ForeignKey(User, verbose_name="Автор", on_delete=models.PROTECT)
     category = models.ForeignKey(Category, verbose_name="Категория", on_delete=models.CASCADE)
     is_deleted = models.BooleanField(verbose_name="Удалена", default=False)
-
-    def save(self, *args, **kwargs):
-        if not self.id:
-            self.created = timezone.now()
-        self.updated = timezone.now()
-        return super().save(*args, **kwargs)
